Split extension into two – the classic one and a new one that requires Manifolds as well.
