+++ Project.toml
@@ -30,7 +30,8 @@ RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
 RipQP = "1e40b3f8-35eb-4cd8-8edd-3e515bb9de08"
 
 [extensions]
-ManoptJuMPExt = "JuMP"
+ManoptJuMPExt = ["JuMP"]
+ManoptJuMPManifoldsExt = ["JuMP", "Manifolds"]
 ManoptLRUCacheExt = "LRUCache"
 ManoptLineSearchesExt = "LineSearches"
 ManoptManifoldsExt = "Manifolds"
